requirements.txt
astroid==1.5.3
dj-database-url==0.4.2
Django==1.11.6
django-colorfield==0.1.14
django-filter==1.0.4
django-filters==0.2.1
django-ordered-model==1.4.3
django-rest-framework==0.1.0
djangorestframework==3.7.0
djangorestframework-filters==0.10.1
gunicorn==19.7.1
isort==4.2.15
lazy-object-proxy==1.3.1
Markdown==2.6.9
mccabe==0.6.1
psycopg2==2.7.3.1
pylint==1.7.4
pytz==2017.2
six==1.11.0
whitenoise==3.3.1
wrapt==1.10.11
